Updates export of RootState and refactors UserInfo slice in features
- This commit updates the export of RootState in index.ts, ensuring consistency.
- It also refactors the UserInfo slice in userInfo.ts for improved clarity and maintainability.
