+++ frontend/src/features/index.ts
@@ -2,4 +2,4 @@
 
 export { store } from './store';
 export type { AppDispatch } from './store';
-export type { RootState } from './rootReducer';
+export type { RootState } from './store';
