+++ frontend/src/features/rootReducer.ts
@@ -8,5 +8,3 @@ const rootReducer = combineReducers({
 })
 
 export default rootReducer;
-
-export type RootState = ReturnType<typeof rootReducer>;
