+++ frontend/src/features/userInfo/userInfoModel.ts
@@ -1,0 +1,7 @@
+type stringOrNull = string | null;
+
+export interface UserInfo {
+  id: stringOrNull;
+  name: stringOrNull;
+  rating: number;
+}
